Migrate metrics to ts
